requirements.txt
Django>=3.2,<6.0
djangorestframework
djangorestframework-simplejwt
django-filter
pyotp
qrcode
Pillow
psycopg2-binary
bleach
django-cors-headers